Move squid components functions hooks to lib folder
