+++ packages/checkout/widgets-lib/src/lib/squid/components/SquidFooter.tsx
