+++ packages/checkout/widgets-lib/src/lib/squid/components/SquidIcon.tsx
