+++ packages/checkout/widgets-lib/src/lib/squid/config.ts
@@ -1,0 +1,3 @@
+export const SQUID_SDK_BASE_URL = 'https://apiplus.squidrouter.com';
+
+export const SQUID_NATIVE_TOKEN = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE';
