+++ packages/checkout/widgets-lib/src/lib/squid/functions/fetchBalances.ts
