+++ packages/checkout/widgets-lib/src/lib/squid/functions/fetchChains.ts
