+++ packages/checkout/widgets-lib/src/lib/squid/functions/getTotalRouteFees.ts
