+++ packages/checkout/widgets-lib/src/lib/squid/functions/sortRoutesByFastestTime.ts
